# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FitTrack API — fitness tracking, nutrition logging, and social features
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /nutrition/foods:
    get:
      operationId: listFoodItems
      tags: [nutrition]
      summary: List food items (recent + custom)
      parameters:
        - name: q
          in: query
          schema:
            type: string
        - name: recentOnly
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Food items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FoodItem"
    post:
      operationId: createFoodItem
      tags: [nutrition]
      summary: Create a custom food item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FoodItemInput"
      responses:
        "201":
          description: Created food item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FoodItem"
  /nutrition/foods/{foodItemId}:
    get:
      operationId: getFoodItem
      tags: [nutrition]
      summary: Get food item details
      parameters:
        - name: foodItemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Food item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FoodItem"
    patch:
      operationId: updateFoodItem
      tags: [nutrition]
      summary: Update a custom food item
      parameters:
        - name: foodItemId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FoodItemUpdate"
      responses:
        "200":
          description: Updated food item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FoodItem"
  /nutrition/logs:
    get:
      operationId: listFoodLogs
      tags: [nutrition]
      summary: List food logs for a date
      parameters:
        - name: date
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Food logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FoodLog"
    post:
      operationId: logFood
      tags: [nutrition]
      summary: Log a food item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FoodLogInput"
      responses:
        "201":
          description: Created food log
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FoodLog"
  /nutrition/logs/{logId}:
    patch:
      operationId: updateFoodLog
      tags: [nutrition]
      summary: Update a food log (quantity, custom macros)
      parameters:
        - name: logId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FoodLogUpdate"
      responses:
        "200":
          description: Updated food log
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FoodLog"
    delete:
      operationId: deleteFoodLog
      tags: [nutrition]
      summary: Delete a food log
      parameters:
        - name: logId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /nutrition/summary:
    get:
      operationId: getNutritionSummary
      tags: [nutrition]
      summary: Get nutrition summary for a date (totals vs goals)
      parameters:
        - name: date
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Nutrition summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NutritionSummary"
  /nutrition/goals:
    get:
      operationId: getNutritionGoals
      tags: [nutrition]
      summary: Get user nutrition goals
      responses:
        "200":
          description: Nutrition goals
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NutritionGoals"
    put:
      operationId: setNutritionGoals
      tags: [nutrition]
      summary: Set user nutrition goals
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NutritionGoalsInput"
      responses:
        "200":
          description: Updated goals
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NutritionGoals"

  # ─── Social ───────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    FoodItem:
      type: object
      required: [id, name, calories, protein, carbs, fats]
      properties:
        id:
          type: integer
        name:
          type: string
        calories:
          type: number
        protein:
          type: number
        carbs:
          type: number
        fats:
          type: number
        healthScore:
          type: ["number", "null"]
        servingSize:
          type: ["number", "null"]
        servingUnit:
          type: ["string", "null"]
        isCustom:
          type: boolean
        usageCount:
          type: integer
    FoodItemInput:
      type: object
      required: [name, calories, protein, carbs, fats]
      properties:
        name:
          type: string
        calories:
          type: number
        protein:
          type: number
        carbs:
          type: number
        fats:
          type: number
        healthScore:
          type: number
        servingSize:
          type: number
        servingUnit:
          type: string
    FoodItemUpdate:
      type: object
      properties:
        name:
          type: string
        calories:
          type: number
        protein:
          type: number
        carbs:
          type: number
        fats:
          type: number
        healthScore:
          type: number
        servingSize:
          type: number
        servingUnit:
          type: string
    FoodLog:
      type: object
      required: [id, foodItemId, foodItem, mealType, quantity, date, loggedAt]
      properties:
        id:
          type: integer
        foodItemId:
          type: integer
        foodItem:
          $ref: "#/components/schemas/FoodItem"
        mealType:
          type: string
          enum: [breakfast, lunch, dinner, snack]
        quantity:
          type: number
        date:
          type: string
          format: date
        loggedAt:
          type: string
          format: date-time
        customCalories:
          type: ["number", "null"]
        customProtein:
          type: ["number", "null"]
        customCarbs:
          type: ["number", "null"]
        customFats:
          type: ["number", "null"]
        effectiveCalories:
          type: number
        effectiveProtein:
          type: number
        effectiveCarbs:
          type: number
        effectiveFats:
          type: number
    FoodLogInput:
      type: object
      required: [foodItemId, mealType, quantity, date]
      properties:
        foodItemId:
          type: integer
        mealType:
          type: string
          enum: [breakfast, lunch, dinner, snack]
        quantity:
          type: number
        date:
          type: string
          format: date
        customCalories:
          type: number
        customProtein:
          type: number
        customCarbs:
          type: number
        customFats:
          type: number
    FoodLogUpdate:
      type: object
      properties:
        quantity:
          type: number
        mealType:
          type: string
          enum: [breakfast, lunch, dinner, snack]
        customCalories:
          type: number
        customProtein:
          type: number
        customCarbs:
          type: number
        customFats:
          type: number
    NutritionSummary:
      type: object
      required: [date, totalCalories, totalProtein, totalCarbs, totalFats]
      properties:
        date:
          type: string
          format: date
        totalCalories:
          type: number
        totalProtein:
          type: number
        totalCarbs:
          type: number
        totalFats:
          type: number
        goalCalories:
          type: ["number", "null"]
        goalProtein:
          type: ["number", "null"]
        goalCarbs:
          type: ["number", "null"]
        goalFats:
          type: ["number", "null"]
        caloriesRemaining:
          type: number
        logs:
          type: array
          items:
            $ref: "#/components/schemas/FoodLog"
    NutritionGoals:
      type: object
      required: [id, calories, protein, carbs, fats]
      properties:
        id:
          type: integer
        calories:
          type: number
        protein:
          type: number
        carbs:
          type: number
        fats:
          type: number
    NutritionGoalsInput:
      type: object
      required: [calories, protein, carbs, fats]
      properties:
        calories:
          type: number
        protein:
          type: number
        carbs:
          type: number
        fats:
          type: number

    # ─── Social ────────────────────────────────────────────────────────────────
...